+++ include/Estatistica.hpp
@@ -9,7 +9,7 @@
 class Estatistica {
 public:
     // Construtor: Recebe uma lista de compras para análise estatística.
-    Estatistica(const std::vector<Compra>& compras, float salario);
+    Estatistica(const std::vector<Compra>& compras, double salario);
 
     // Calcula os gastos por categoria e armazena no mapa.
     void calcularGastosPorCategoria();         
@@ -18,20 +18,20 @@ public:
     void exibirEstatisticas();                 
 
     // Analisa os gastos com base no salário do usuário.
-    void analisarComBaseSalario(float salario) const; 
+    void analisarComBaseSalario(double salario) const; 
 
     // Exibe uma comparação dos gastos com o salário informado.
-    void exibirComparacaoComSalario(float salario) const; 
+    void exibirComparacaoComSalario(double salario) const; 
 
 private:
     const std::vector<Compra> compras;               // Lista de compras analisadas (imutável)
-    std::map<std::string, float> gastosPorCategoria; // Mapa de categoria -> valor gasto
-    float totalGasto;                                // Total gasto calculado
+    std::map<std::string, double> gastosPorCategoria; // Mapa de categoria -> valor gasto
+    double totalGasto;                                // Total gasto calculado
 
     // Calcula o total gasto somando os valores de todas as compras.
     void calcularTotalGasto();      
 
-    float salarioUsuario;        
+    double salarioUsuario;        
 };
 
 #endif
