+++ src/Estatistica.cpp
@@ -4,7 +4,7 @@
 #include "Sistema.hpp"
 #include "Operacoes.hpp"
 
-Estatistica::Estatistica(const std::vector<Compra>& compras, float salario) 
+Estatistica::Estatistica(const std::vector<Compra>& compras, double salario) 
     : compras(compras), totalGasto(0.0f), salarioUsuario(salario) {}
 
 void Estatistica::calcularTotalGasto() {
@@ -14,7 +14,7 @@ void Estatistica::calcularTotalGasto() {
     }
 }
 
-void Estatistica::analisarComBaseSalario(float salario) const {
+void Estatistica::analisarComBaseSalario(double salario) const {
     if (gastosPorCategoria.empty() || salario <= 0) {
         std::cout << "Nenhum dado para análise ou salário inválido." << std::endl;
         return;
@@ -22,7 +22,7 @@ void Estatistica::analisarComBaseSalario(float salario) const {
 
     std::cout << "--- Comparação com Salário ---" << std::endl;
     for (const auto& [categoria, valor] : gastosPorCategoria) {
-        float porcentagemSalario = (valor / salario) * 100;
+        double porcentagemSalario = (valor / salario) * 100;
         std::cout << "Categoria: " << categoria 
                   << " - " << std::fixed << std::setprecision(2) << porcentagemSalario
                   << "% do salário" << std::endl;
@@ -36,7 +36,7 @@ void Estatistica::exibirEstatisticas() {
     std::cout << "Total gasto: R$ " << totalGasto << "\n";
 
     for (const auto& [categoria, valor] : gastosPorCategoria) {
-        float porcentagem = (valor / totalGasto) * 100;
+        double porcentagem = (valor / totalGasto) * 100;
         std::cout << "Categoria: " << categoria
                   << " - R$ " << valor
                   << " (" << std::fixed << std::setprecision(2) << porcentagem << "%)\n";
@@ -44,7 +44,7 @@ void Estatistica::exibirEstatisticas() {
 
     std::cout << "\n--- Comparação com Salário ---\n";
     for (const auto& [categoria, valor] : gastosPorCategoria) {
-        float porcentagemSalario = (valor / salarioUsuario) * 100;
+        double porcentagemSalario = (valor / salarioUsuario) * 100;
         std::cout << "Categoria: " << categoria 
                   << " - " << std::fixed << std::setprecision(2) << porcentagemSalario
                   << "% do salário" << std::endl;
@@ -52,10 +52,10 @@ void Estatistica::exibirEstatisticas() {
 }
 
 
-void Estatistica::exibirComparacaoComSalario(float salario) const {
+void Estatistica::exibirComparacaoComSalario(double salario) const {
     std::cout << "--- Comparação com Salário ---" << std::endl;
     for (const auto& [categoria, valor] : gastosPorCategoria) {
-        float porcentagemSalario = (valor / salario) * 100;
+        double porcentagemSalario = (valor / salario) * 100;
         std::cout << "Categoria: " << categoria 
                   << " - " << std::fixed << std::setprecision(2) << porcentagemSalario
                   << "% do salário" << std::endl;
@@ -67,7 +67,7 @@ void Estatistica::calcularGastosPorCategoria() {
     totalGasto = 0.0f;          // Reinicia o total
 
     for (const auto& compra : compras) {
-        float valor = compra.getValor();
+        double valor = compra.getValor();
         std::string categoria = compra.getCategoria();
 
         // Soma ao total e atualiza a categoria
